Switch mixed precision from fp16 to bf16
Move both the forward compute dtype and the FSDP parameter/reduce dtypes from
fp16 to bf16 to eliminate the fp16 overflow path. Changing only one would leave
overflow reachable, so both are flipped together.

- dinov2/train/train.py: inputs_dtype torch.half -> torch.bfloat16 (forward
  compute dtype passed to the collate cast).
- dinov2/configs/ssl_default_config.yaml: within compute_precision, flip every
  param_dtype/reduce_dtype fp16 -> bf16. buffer_dtype stays fp32 (norm buffers
  in fp32 for stability), and the student head reduce_dtype entries that were
  already fp32 are left unchanged. bf16 is already in the FSDP dtype map.

The grad scaler is config-controlled (ssl_meta_arch reads
compute_precision.grad_scaler) and bf16 needs no loss scaling; the default is
left as-is and disabled per-run in the generated run config.

+++ dinov2/train/train.py
@@ -252,7 +252,7 @@ def do_test(cfg, model, iteration):
 
 def do_train(cfg, model, resume=False):
     model.train()
-    inputs_dtype = torch.half
+    inputs_dtype = torch.bfloat16  # was: torch.half — bf16 forward compute avoids fp16 overflow
     fp16_scaler = model.fp16_scaler  # for mixed precision training
 
     # setup optimizer
